+++ tests/api-create.js
@@ -1,8 +1,12 @@
 'use strict';
 
+require('lllog')('none');
+
 const APITest = require('@janiscommerce/api-test');
-const { Invoker } = require('@janiscommerce/lambda');
 const Settings = require('@janiscommerce/settings');
+const MicroserviceCall = require('@janiscommerce/microservice-call');
+
+const { Invoker } = require('@janiscommerce/lambda');
 
 const { APICreate, ModelClient } = require('../lib');
 
@@ -31,6 +35,7 @@ describe('Client Create API', () => {
 	const clients = ['foo', 'bar'];
 
 	const clientsToSave = clients.map(code => prepareFakeClient(code));
+	const clientsToSaveWithAdditionalFields = clientsToSave.map(client => ({ ...client, extraField: 'some-data' }));
 
 	const janisServiceName = 'some-service-name';
 	process.env.JANIS_SERVICE_NAME = janisServiceName;
@@ -54,6 +59,10 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -67,13 +76,64 @@ describe('Client Create API', () => {
 
 				assertSecretsGet(sandbox, janisServiceName);
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
-
 				sandbox.assert.calledOnceWithExactly(Invoker.call, 'MongoDBIndexCreator');
 
 				sandbox.assert.calledOnceWithExactly(
 					APICreate.prototype.postSaveHook,
-					clients
+					clients,
+					clientsToSave
+				);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should save all the received new clients to clients DB including additional fields',
+			request: {
+				data: { clients }
+			},
+			response: { code: 200 },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(ModelClient, 'additionalFields')
+					.get(() => ['extraField']);
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSaveWithAdditionalFields });
+
+				sandbox.stub(ModelClient.prototype, 'multiSave')
+					.resolves(true);
+
+				sandbox.stub(Invoker, 'call')
+					.resolves();
+
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				assertSecretsGet(sandbox, janisServiceName);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSaveWithAdditionalFields);
+				sandbox.assert.calledOnceWithExactly(Invoker.call, 'MongoDBIndexCreator');
+
+				sandbox.assert.calledOnceWithExactly(
+					APICreate.prototype.postSaveHook,
+					clients,
+					clientsToSaveWithAdditionalFields
 				);
 
 				stopMock();
@@ -97,6 +157,10 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -110,13 +174,16 @@ describe('Client Create API', () => {
 
 				secretsNotCalled(sandbox);
 
-				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clients.map(code => prepareFakeClient(code, false, false)));
+				const expectedClientsToSave = clients.map(code => prepareFakeClient(code, false, false));
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, expectedClientsToSave);
 				sandbox.assert.calledOnceWithExactly(Invoker.call, 'MongoDBIndexCreator');
 
 				sandbox.assert.calledOnceWithExactly(
 					APICreate.prototype.postSaveHook,
-					clients
+					clients,
+					expectedClientsToSave
 				);
 
 				stopMock();
@@ -149,6 +216,10 @@ describe('Client Create API', () => {
 					}
 				});
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -162,6 +233,7 @@ describe('Client Create API', () => {
 
 				assertSecretsGet(sandbox, janisServiceName);
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clients.map(code => prepareFakeClient(code, true)));
 
 				stopMock();
@@ -184,6 +256,10 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -197,6 +273,7 @@ describe('Client Create API', () => {
 
 				secretsNotCalled(sandbox);
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
 
 				stopMock();
@@ -219,6 +296,10 @@ describe('Client Create API', () => {
 
 				secretThrows(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -230,6 +311,7 @@ describe('Client Create API', () => {
 			},
 			after: (res, sandbox) => {
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
 
 				stopMock();
@@ -252,6 +334,10 @@ describe('Client Create API', () => {
 
 				getValueRejects(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves(true);
 
@@ -263,7 +349,207 @@ describe('Client Create API', () => {
 			},
 			after: (res, sandbox) => {
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should return 200 when MicroserviceCall can\'t get clients from ID Service',
+			request: {
+				data: { clients }
+			},
+			response: { code: 200 },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 400, body: {} });
+
+				sandbox.spy(ModelClient.prototype, 'multiSave');
+				sandbox.spy(Invoker, 'call');
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				secretsNotCalled(sandbox);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.notCalled(ModelClient.prototype.multiSave);
+				sandbox.assert.notCalled(Invoker.call);
+				sandbox.assert.notCalled(APICreate.prototype.postSaveHook);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should save only found clients when can\'t get all clients from ID service',
+			request: {
+				data: { clients }
+			},
+			response: { code: 200 },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: [clientsToSave[0]] });
+
+				sandbox.stub(ModelClient.prototype, 'multiSave')
+					.resolves(true);
+
+				sandbox.stub(Invoker, 'call')
+					.resolves();
+
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				assertSecretsGet(sandbox, janisServiceName);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, [clientsToSave[0]]);
+				sandbox.assert.calledOnceWithExactly(Invoker.call, 'MongoDBIndexCreator');
+
+				sandbox.assert.calledOnceWithExactly(
+					APICreate.prototype.postSaveHook,
+					clients,
+					[clientsToSave[0]]
+				);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should return 200 when MicroserviceCall don\'t get any clients from ID Service',
+			request: {
+				data: { clients }
+			},
+			response: { code: 200 },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: [] });
+
+				sandbox.spy(ModelClient.prototype, 'multiSave');
+				sandbox.spy(Invoker, 'call');
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				secretsNotCalled(sandbox);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.notCalled(ModelClient.prototype.multiSave);
+				sandbox.assert.notCalled(Invoker.call);
+				sandbox.assert.notCalled(APICreate.prototype.postSaveHook);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should return 500 when MicroserviceCall fails to get clients from ID Service',
+			request: {
+				data: { clients }
+			},
+			response: { code: 500, body: { message: 'Failed to get Janis ID clients: Service failed' } },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 500, body: {} });
+
+				sandbox.spy(ModelClient.prototype, 'multiSave');
+				sandbox.spy(Invoker, 'call');
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				secretsNotCalled(sandbox);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.notCalled(ModelClient.prototype.multiSave);
+				sandbox.assert.notCalled(Invoker.call);
+				sandbox.assert.notCalled(APICreate.prototype.postSaveHook);
+
+				stopMock();
+			}
+		}, {
+			description: 'Should return 500 when MicroserviceCall fails to get clients from ID Service with API Error message',
+			request: {
+				data: { clients }
+			},
+			response: { code: 500, body: { message: 'Failed to get Janis ID clients: Some API Error' } },
+			before: sandbox => {
+
+				delete ClientFormatter.settings;
+
+				mockModelClient();
+
+				sandbox.stub(Settings, 'get').returns(fakeDBSettings);
+
+				setEnv();
+
+				stubGetSecret(sandbox);
+
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 500, body: { message: 'Some API Error' } });
+
+				sandbox.spy(ModelClient.prototype, 'multiSave');
+				sandbox.spy(Invoker, 'call');
+				sandbox.spy(APICreate.prototype, 'postSaveHook');
+
+			},
+			after: (res, sandbox) => {
+
+				secretsNotCalled(sandbox);
+
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
+				sandbox.assert.notCalled(ModelClient.prototype.multiSave);
+				sandbox.assert.notCalled(Invoker.call);
+				sandbox.assert.notCalled(APICreate.prototype.postSaveHook);
 
 				stopMock();
 			}
@@ -285,6 +571,10 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.rejects();
 
@@ -295,6 +585,7 @@ describe('Client Create API', () => {
 
 				assertSecretsGet(sandbox, janisServiceName);
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
 				sandbox.assert.notCalled(Invoker.call);
 
@@ -318,6 +609,10 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
+				sandbox.stub(MicroserviceCall.prototype, 'safeList')
+					.withArgs('id', 'client', { filters: { clientCode: clients } })
+					.resolves({ statusCode: 200, body: clientsToSave });
+
 				sandbox.stub(ModelClient.prototype, 'multiSave')
 					.resolves();
 
@@ -328,9 +623,10 @@ describe('Client Create API', () => {
 
 				assertSecretsGet(sandbox, janisServiceName);
 
+				sandbox.assert.calledOnceWithExactly(MicroserviceCall.prototype.safeList, 'id', 'client', { filters: { clientCode: clients } });
 				sandbox.assert.calledOnceWithExactly(ModelClient.prototype.multiSave, clientsToSave);
-
 				sandbox.assert.calledOnceWithExactly(Invoker.call, 'MongoDBIndexCreator');
+
 				stopMock();
 			}
 		}, {
@@ -373,20 +669,19 @@ describe('Client Create API', () => {
 
 				stubGetSecret(sandbox);
 
-				sandbox.stub(ModelClient.prototype, 'multiSave')
-					.resolves(true);
-
-				sandbox.stub(Invoker, 'call')
-					.resolves();
+				sandbox.spy(MicroserviceCall.prototype, 'safeList');
+				sandbox.spy(ModelClient.prototype, 'multiSave');
+				sandbox.spy(Invoker, 'call');
 			},
 			after: (res, sandbox) => {
 
+				sandbox.assert.notCalled(MicroserviceCall.prototype.safeList);
 				sandbox.assert.notCalled(ModelClient.prototype.multiSave);
 				sandbox.assert.notCalled(Invoker.call);
 
 				stopMock();
 
-				assertSecretsGet(sandbox, janisServiceName);
+				secretsNotCalled(sandbox);
 			}
 		}
 	]);